example/requirements.txt
deep-trainer
torch
torch-resnet
torchvision
pyyaml
